initial commit of these project

+++ F429_sensor/Core/Src/retarget.c
@@ -1,14 +1,13 @@
-#include "main.h"
-#include <stdio.h>
-#include <stdint.h>
+#include "stm32f4xx_hal.h"
+#include <sys/unistd.h>
 #include "stm32f4xx_hal.h"
 #include "stm32f4xx_hal_uart.h"
+#include <stdint.h>
 
 extern UART_HandleTypeDef huart3;
 
 int _write(int file, char *ptr, int len)
 {
-    HAL_UART_Transmit(&huart3 , (uint8_t *)ptr , (uint16_t)len , HAL_MAX_DELAY);
+    HAL_UART_Transmit(&huart3 , (uint8_t *)ptr, len , HAL_MAX_DELAY);
     return len;
-} // end _write
-
+}
